# Improve performance of bit-parallel counting methods