--- a/src/bit-manipulation/bitpar-compiler.jl
+++ b/src/bit-manipulation/bitpar-compiler.jl
@@ -28,14 +28,8 @@ function compile_bitpar(funcname::Symbol;
                 chunk = (data[index(ind)] >> o) & bitmask(stop - ind)
                 $(head_code)
                 ind += 64 - o
-                @assert offset(ind) == 0
+                #@assert offset(ind) == 0
             end
-            
-            #while i ≤ stop - 64
-            #    chunk = seq.data[index(i)]
-            #    $(body_code)
-            #    i += 64
-            #end
             
             lastind = index(stop - bits_per_symbol(seq))
             lastind -= !iszero(offset(stop))