This PR aims to fix some of the observations from the performance review #86 

Notably:

- [x] Slow GC counting in bit parallel mode.
- [x] Lack of bit-parallel single sequence counting mode for ambiguous nucleotides.


Benchmarks prior to changes:

```julia
julia> using BioSequences, BenchmarkTools
julia> s = randdnaseq(100_000)
julia> s2 = LongSequence{DNAAlphabet{2}}(s)

julia> @benchmark count(isGC, s)
BenchmarkTools.Trial: 
  memory estimate:  16 bytes
  allocs estimate:  1
  --------------
  minimum time:     8.385 μs (0.00% GC)
  median time:      8.499 μs (0.00% GC)
  mean time:        8.784 μs (0.00% GC)
  maximum time:     29.168 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     3

julia> @benchmark count(isGC, s2)
BenchmarkTools.Trial: 
  memory estimate:  16 bytes
  allocs estimate:  1
  --------------
  minimum time:     2.723 μs (0.00% GC)
  median time:      2.746 μs (0.00% GC)
  mean time:        2.829 μs (0.00% GC)
  maximum time:     8.697 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     9

julia> @benchmark count(isambiguous, s)
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     114.989 μs (0.00% GC)
  median time:      119.960 μs (0.00% GC)
  mean time:        136.786 μs (0.00% GC)
  maximum time:     413.966 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

julia> @benchmark count(isambiguous, s2)
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     133.609 μs (0.00% GC)
  median time:      136.452 μs (0.00% GC)
  mean time:        144.586 μs (0.00% GC)
  maximum time:     407.740 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```