@BenJWard 2020-01-27
See conversation in #86. This should improve the performance of bit-parallel GC counting.