@BenJWard 2020-01-27
@jakobnissen The performance of `count(isGC, seq)` on this branch should now be comparable with your `simple_gc` version. Can you verify this works for you and isn't specific to code compiled on my machine?