@BenJWard 2020-01-27
Benchmarks after changes:

```julia
julia> @benchmark count(isGC, s)
BenchmarkTools.Trial: 
  memory estimate:  16 bytes
  allocs estimate:  1
  --------------
  minimum time:     2.960 μs (0.00% GC)
  median time:      3.015 μs (0.00% GC)
  mean time:        3.050 μs (0.00% GC)
  maximum time:     10.778 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     9

julia> @benchmark count(isGC, s2)
BenchmarkTools.Trial: 
  memory estimate:  16 bytes
  allocs estimate:  1
  --------------
  minimum time:     929.063 ns (0.00% GC)
  median time:      960.656 ns (0.00% GC)
  mean time:        992.586 ns (0.00% GC)
  maximum time:     4.031 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     32
```
```