@jakobnissen 2020-01-29
This all looks good to me! I get the same timings on my computer with your updated code as my `simple_gc`. Also, I get similar timings for `n_ambiguous`, and verify that `n_ambiguous` for 2-bit nucleotides is instant.
The ambiguous bitcounters still need some documentation, but I can make a PR for that later, if you don't mind waiting for up to a few weeks.