@BenJWard 2020-02-11
I'll merge this rather than let it hang. There's still probably some improvements to make for the 2 sequence counting methods, and docs, but we can have that as another PR.